refactor negotiator
