+++ src/groups/mqb/mqba/package/mqba.mem
@@ -6,4 +6,5 @@ mqba_configprovider
 mqba_dispatcher
 mqba_domainmanager
 mqba_domainresolver
+mqba_negotiationcontext
 mqba_sessionnegotiator
