+++ src/groups/mqb/mqbnet/mqbnet_negotiator.h
@@ -21,7 +21,6 @@
 //
 //@CLASSES:
 //  mqbnet::Negotiator:        protocol for a session negotiator
-//  mqbnet::NegotiatorContext: VST for the context associated to a negotiation
 //
 //@SEE_ALSO:
 //  mqbnet::Session: protocol of the negotiated sessions
@@ -44,6 +43,7 @@
 // of its members that the caller will leverage and use.
 
 // MQB
+#include <mqbnet_negotiatorcontext.h>
 
 // BDE
 #include <bsl_functional.h>
@@ -64,109 +64,6 @@ namespace mqbnet {
 class Session;
 class SessionEventProcessor;
 class Cluster;
-
-// =======================
-// class NegotiatorContext
-// =======================
-
-/// VST for the context associated to a session being negotiated.  Each
-/// session being negotiated get its own context; and the Negotiator
-/// concrete implementation can modify some of the members during the
-/// negotiation (i.e., between the `negotiate()` method and the invocation
-/// of the `NegotiationCb` method.
-class NegotiatorContext {
-  private:
-    // DATA
-    bool d_isIncoming;
-    // True if the session being negotiated originates
-    // from a remote peer (i.e., a 'listen'); false if
-    // it originates from us (i.e., a 'connect).
-
-    int d_maxMissedHeartbeat;
-    // If non-zero, enable smart-heartbeat and specify
-    // that the connection should be proactively
-    // resetted if no data has been received from this
-    // channel for the 'maxMissedHeartbeat' number of
-    // heartbeat intervals.  When enabled, heartbeat
-    // requests will be sent if no 'regular' data is
-    // being received.
-
-    SessionEventProcessor* d_eventProcessor_p;
-    // The event processor to use for initiating the
-    // read on the channel once the session has been
-    // successfully negotiated.  This may or may not be
-    // set by the caller, before invoking
-    // 'Negotiator::negotiate()'; and may or may not be
-    // changed by the negotiator concrete
-    // implementation before invoking the
-    // 'NegotiationCb'.  Note that a value of 0 will
-    // use the negotiated session as the default event
-    // processor.
-
-    void* d_resultState_p;
-    // Raw pointer, held not owned, to some user data
-    // the session factory will pass back to the
-    // 'resultCb' method (used to inform of the
-    // success/failure of a session negotiation).  This
-    // may or may not be set by the caller, before
-    // invoking 'Negotiator::negotiate()'; and may or
-    // may not be changed by the negotiator concrete
-    // implementation before invoking the
-    // 'NegotiationCb'.  This is used to bind low level
-    // data (from transport layer) to the session; and
-    // can be overriden/set by the negotiation
-    // implementation (typically for the case of
-    // 'listen' sessions, since those are
-    // 'sporadically' happening and there is not enough
-    // context at the transport layer to find back this
-    // data).
-
-    void* d_userData_p;
-    // Raw pointer, held not owned, to some user data
-    // the Negotiator concrete implementation can use
-    // while negotiating the session.  This may or may
-    // not be set by the caller, before invoking
-    // 'Negotiator::negotiate()'; and should not be
-    // changed during negotiation (this data is not
-    // used by the session factory, so changing it will
-    // have no effect).  This is used to bind high
-    // level data (from application layer) to the
-    // application layer (the negotiator concrete
-    // implementation) (typically for the case of
-    // 'connect' sessions to provide information to use
-    // for negotiating the session with the remote
-    // peer).
-
-    Cluster* d_cluster_p;
-    // mqbnet::Cluster to inform about incoming (proxy)
-    // connection
-
-  public:
-    // CREATORS
-
-    /// Create a new object having the specified `isIncoming` value.
-    NegotiatorContext(bool isIncoming);
-
-    // MANIPULATORS
-    NegotiatorContext& setMaxMissedHeartbeat(int value);
-    NegotiatorContext& setUserData(void* value);
-    NegotiatorContext& setResultState(void* value);
-    NegotiatorContext& setEventProcessor(SessionEventProcessor* value);
-
-    /// Set the corresponding field to the specified `value` and return a
-    /// reference offering modifiable access to this object.
-    NegotiatorContext& setCluster(Cluster* cluster);
-
-    // ACCESSORS
-    bool     isIncoming() const;
-    Cluster* cluster() const;
-    int      maxMissedHeartbeat() const;
-    void*    userData() const;
-    void*    resultState() const;
-
-    /// Return the value of the corresponding field.
-    SessionEventProcessor* eventProcessor() const;
-};
 
 // ================
 // class Negotiator
@@ -212,86 +109,6 @@ class Negotiator {
                            const bsl::shared_ptr<bmqio::Channel>& channel,
                            const NegotiationCb& negotiationCb) = 0;
 };
-
-// ============================================================================
-//                             INLINE DEFINITIONS
-// ============================================================================
-
-// -----------------------
-// class NegotiatorContext
-// -----------------------
-
-inline NegotiatorContext::NegotiatorContext(bool isIncoming)
-: d_isIncoming(isIncoming)
-, d_maxMissedHeartbeat(0)
-, d_eventProcessor_p(0)
-, d_resultState_p(0)
-, d_userData_p(0)
-, d_cluster_p(0)
-{
-    // NOTHING
-}
-
-inline NegotiatorContext& NegotiatorContext::setMaxMissedHeartbeat(int value)
-{
-    d_maxMissedHeartbeat = value;
-    return *this;
-}
-
-inline NegotiatorContext& NegotiatorContext::setUserData(void* value)
-{
-    d_userData_p = value;
-    return *this;
-}
-
-inline NegotiatorContext& NegotiatorContext::setResultState(void* value)
-{
-    d_resultState_p = value;
-    return *this;
-}
-
-inline NegotiatorContext&
-NegotiatorContext::setEventProcessor(SessionEventProcessor* value)
-{
-    d_eventProcessor_p = value;
-    return *this;
-}
-
-inline NegotiatorContext& NegotiatorContext::setCluster(Cluster* cluster)
-{
-    d_cluster_p = cluster;
-    return *this;
-}
-
-inline bool NegotiatorContext::isIncoming() const
-{
-    return d_isIncoming;
-}
-
-inline Cluster* NegotiatorContext::cluster() const
-{
-    return d_cluster_p;
-}
-
-inline int NegotiatorContext::maxMissedHeartbeat() const
-{
-    return d_maxMissedHeartbeat;
-}
-
-inline void* NegotiatorContext::userData() const
-{
-    return d_userData_p;
-}
-
-inline void* NegotiatorContext::resultState() const
-{
-    return d_resultState_p;
-}
-
-inline SessionEventProcessor* NegotiatorContext::eventProcessor() const
-{
-    return d_eventProcessor_p;
-}
 
 }  // close package namespace
 }  // close enterprise namespace
