+++ src/groups/mqb/mqbnet/package/mqbnet.mem
@@ -6,6 +6,7 @@ mqbnet_dummysession
 mqbnet_mockcluster
 mqbnet_multirequestmanager
 mqbnet_negotiator
+mqbnet_negotiatorcontext
 mqbnet_session
 mqbnet_tcpsessionfactory
 mqbnet_transportmanager
